Move book list page component (#178)
