+++ frontend/app/routes/home._index/components/BookCard.tsx
