+++ frontend/app/routes/home._index/components/BookCardFooter.tsx
