+++ frontend/app/routes/home._index/components/BookCardHeader.tsx
