+++ frontend/app/routes/home._index/components/BookCardThumbnail.tsx
